input tests with placeholer
